+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Defunc.hs
@@ -7,7 +7,7 @@ import Parsley.Internal.Backend.Machine.InputRep (Rep)
 import Parsley.Internal.Common.Utils             (Code)
 import Parsley.Internal.Core.Lam                 (Lam, normaliseGen, normalise)
 
-import qualified Parsley.Internal.Core.Defunc as Core (Defunc, lamTerm)
+import qualified Parsley.Internal.Core.Defunc as Core (Defunc, lamTerm, lamTermBool)
 import qualified Parsley.Internal.Core.Lam    as Lam  (Lam(..))
 
 data Defunc a where
@@ -19,27 +19,32 @@ data Defunc a where
 user :: Core.Defunc a -> Defunc a
 user = LAM . Core.lamTerm
 
+userBool :: Core.Defunc (a -> Bool) -> Defunc (a -> Bool)
+userBool = LAM . Core.lamTermBool
+
+ap :: Defunc (a -> b) -> Defunc a -> Defunc b
+ap f x = LAM (Lam.App (unliftDefunc f) (unliftDefunc x))
+
 ap2 :: Defunc (a -> b -> c) -> Defunc a -> Defunc b -> Defunc c
 ap2 f@SAME (OFFSET o1) (OFFSET o2) = LAM (Lam.Var False (apSame f o1 o2))
   where
     apSame :: forall o. Defunc (o -> o -> Bool) -> Code (Rep o) -> Code (Rep o) -> Code Bool
     apSame SAME = same (Proxy @o)
     apSame _    = undefined
-ap2 f x y = LAM (Lam.App (Lam.App (seal f) (seal x)) (seal y))
-  where
-    seal :: Defunc a -> Lam a
-    seal (LAM x) = x
-    seal x        = Lam.Var False (genDefunc x)
+ap2 f x y = ap (ap f x) y
+
+_if :: Defunc Bool -> Code a -> Code a -> Code a
+_if c t e = normaliseGen (Lam.If (unliftDefunc c) (Lam.Var False t) (Lam.Var False e))
+
+unliftDefunc :: Defunc a -> Lam a
+unliftDefunc (LAM x) = x
+unliftDefunc x       = Lam.Var False (genDefunc x)
 
 genDefunc :: Defunc a -> Code a
 genDefunc (LAM x)    = normaliseGen x
 genDefunc BOTTOM      = [||undefined||]
 genDefunc SAME        = error "Cannot materialise the same function in the regular way"
 genDefunc (OFFSET _)  = error "Cannot materialise an unboxed offset in the regular way"
-
-genDefunc1 :: Defunc (a -> b) -> Code a -> Code b
-genDefunc1 (LAM f) qx = normaliseGen (Lam.App f (Lam.Var True qx))
-genDefunc1 f qx       = [|| $$(genDefunc f) $$qx ||]
 
 pattern NormLam :: Lam a -> Defunc a
 pattern NormLam t <- LAM (normalise -> t)